Remove NTP Header Only
